#447: Make references to SyncAction weak, to prevent memory leak
For some reason, on log out, we're capturing all kinds of guff in some
closures. I don't understand why. However, as a result of that, we
ended up with two SyncEngine objects.

By making references to SyncAction weak, we no longer end up with
multiple syncengines. It doesn't solve the underlying problem, but it
prevents updating progress on the logged out user...