Remove 'Install SwiftLint' step
SwiftLint is installed in GitHub's base virtual environment image since
[`macos-10.14`](https://github.com/actions/virtual-environments/blob/main/images/macos/macos-10.14-Readme.md).

Attempting to install SwiftLint using `brew` was causing this workflow
to fail with the following errors, since `brew` was unable to overwrite
the `swiftlint` symlink already installed in the image:

```
Error: The `brew link` step did not complete successfully
...
Error: Process completed with exit code 1.
```

Note: this used to work because the workflow is using `macos-latest`.
This used to resolve to
[`macos-10.13`](https://github.com/actions/virtual-environments/blob/main/images/macos/macos-10.13-Readme.md)
which didn't include SwiftLint in the base image.